requirements.txt
bs4==0.0.1
requests==2.22.0
pytube
pydantic
PyYAML
python-dotenv